olympia/requirements.txt
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
Fiona==1.9.0
galois==0.3.3
geographiclib==2.0
geopandas==0.12.2
geopy==2.3.0
importlib-metadata==6.0.0
llvmlite==0.39.1
munch==2.5.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
packaging==23.0
pandas==1.5.3
pick==2.2.0
pycparser==2.21
PyNaCl==1.5.0
pyproj==3.4.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
scipy==1.10.0
shapely==2.0.1
six==1.16.0
typing_extensions==4.4.0
#windows-curses==2.3.1
zipp==3.12.1
gmpy2
pycryptodome
